# Hedging now works without routing configuration