--- a/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpClientFactory.cs
+++ b/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpClientFactory.cs
@@ -35,14 +35,13 @@ internal static class HttpClientFactory
     internal const string PrimaryEndpoint = "http://localhost1";
     internal const string SecondaryEndpoint = "http://localhost2";
 
-    public static ServiceProvider InitializeServiceProvider(HedgingClientType clientType)
+    public static ServiceProvider InitializeServiceProvider(params HedgingClientType[] clientType)
     {
         var services = new ServiceCollection();
         services
             .RegisterMetering()
             .AddSingleton<IRedactorProvider>(NullRedactorProvider.Instance)
             .AddTransient<NoRemoteCallHandler>()
-            .AddHedging(clientType)
             .AddHttpClient(StandardClient, client => client.Timeout = Timeout.InfiniteTimeSpan)
             .AddStandardResilienceHandler()
             .Services
@@ -59,10 +58,15 @@ internal static class HttpClientFactory
         services.RemoveAll<ILoggerFactory>();
         services.AddSingleton<ILoggerFactory>(NullLoggerFactory.Instance);
 
+        foreach (var type in clientType)
+        {
+            services.AddHedging(type);
+        }
+
         return services.BuildServiceProvider();
     }
 
-    private static IServiceCollection AddHedging(this IServiceCollection services, HedgingClientType clientType)
+    private static void AddHedging(this IServiceCollection services, HedgingClientType clientType)
     {
         var clientBuilder = services.AddHttpClient(clientType.ToString(), client => client.Timeout = Timeout.InfiniteTimeSpan);
         var hedgingBuilder = clientBuilder.AddStandardHedgingHandler().SelectStrategyByAuthority(SimpleClassifications.PublicData);
@@ -70,7 +74,7 @@ internal static class HttpClientFactory
 
         if (clientType.HasFlag(HedgingClientType.NoRoutes))
         {
-            return services;
+            return;
         }
 
         int routes = clientType.HasFlag(HedgingClientType.ManyRoutes) ? 50 : 2;
@@ -121,7 +125,5 @@ internal static class HttpClientFactory
                 }).ToArray();
             });
         }
-
-        return services;
     }
 }