--- a/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpResilienceBenchmark.cs
+++ b/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpResilienceBenchmark.cs
@@ -12,7 +12,7 @@ namespace Microsoft.Extensions.Http.Resilience.Bench;
 
 public class HttpResilienceBenchmark
 {
-    private static readonly Uri _uri = new("https://bogus");
+    private static readonly Uri _uri = new(HttpClientFactory.PrimaryEndpoint);
 
     private HttpClient _client = null!;
     private HttpClient _standardClient = null!;