--- a/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpResilienceBenchmark.cs
+++ b/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/HttpResilienceBenchmark.cs
@@ -10,6 +10,8 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace Microsoft.Extensions.Http.Resilience.Bench;
 
+#pragma warning disable R9A044 // Assign array of literal values to a static field for improved performance
+
 public class HttpResilienceBenchmark
 {
     private static readonly Uri _uri = new(HttpClientFactory.PrimaryEndpoint);
@@ -17,7 +19,8 @@ public class HttpResilienceBenchmark
     private HttpClient _client = null!;
     private HttpClient _standardClient = null!;
     private HttpClient _singleHandlerClient = null!;
-    private HttpClient _hedgingClient = null!;
+    private HttpClient _hedgingClientOrdered = null!;
+    private HttpClient _hedgingClientNoRoutes = null!;
 
     private static HttpRequestMessage Request
     {
@@ -32,12 +35,13 @@ public class HttpResilienceBenchmark
     [GlobalSetup]
     public void GlobalSetup()
     {
-        var serviceProvider = HttpClientFactory.InitializeServiceProvider(HedgingClientType.Ordered);
+        var serviceProvider = HttpClientFactory.InitializeServiceProvider(HedgingClientType.Ordered, HedgingClientType.NoRoutes);
         var factory = serviceProvider.GetRequiredService<IHttpClientFactory>();
         _client = factory.CreateClient(HttpClientFactory.EmptyClient);
         _standardClient = factory.CreateClient(HttpClientFactory.StandardClient);
         _singleHandlerClient = factory.CreateClient(HttpClientFactory.SingleHandlerClient);
-        _hedgingClient = factory.CreateClient(nameof(HedgingClientType.Ordered));
+        _hedgingClientNoRoutes = factory.CreateClient(nameof(HedgingClientType.NoRoutes));
+        _hedgingClientOrdered = factory.CreateClient(nameof(HedgingClientType.Ordered));
     }
 
     [Benchmark(Baseline = true)]
@@ -47,20 +51,26 @@ public class HttpResilienceBenchmark
     }
 
     [Benchmark]
+    public Task<HttpResponseMessage> SingleHandler()
+    {
+        return _singleHandlerClient.SendAsync(Request, CancellationToken.None);
+    }
+
+    [Benchmark]
     public Task<HttpResponseMessage> StandardResilienceHandler()
     {
         return _standardClient.SendAsync(Request, CancellationToken.None);
     }
 
     [Benchmark]
-    public Task<HttpResponseMessage> StandardHedgingHandler()
+    public Task<HttpResponseMessage> StandardHedgingHandler_RoutesFromRequest()
     {
-        return _hedgingClient.SendAsync(Request, CancellationToken.None);
+        return _hedgingClientNoRoutes.SendAsync(Request, CancellationToken.None);
     }
 
     [Benchmark]
-    public Task<HttpResponseMessage> SingleHandler()
+    public Task<HttpResponseMessage> StandardHedgingHandler_RoutesFromConfig()
     {
-        return _singleHandlerClient.SendAsync(Request, CancellationToken.None);
+        return _hedgingClientOrdered.SendAsync(Request, CancellationToken.None);
     }
 }