--- a/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/Program.cs
+++ b/bench/Libraries/Microsoft.Extensions.Http.Resilience.PerformanceTests/Program.cs
@@ -15,7 +15,7 @@ namespace Microsoft.Extensions.Http.Resilience.FaultInjection.Benchmark
         {
             var doNotRequireSlnToRunBenchmarks = ManualConfig
                 .Create(DefaultConfig.Instance)
-                .AddJob(Job.MediumRun.WithToolchain(InProcessEmitToolchain.Instance).WithIterationCount(100))
+                .AddJob(Job.MediumRun.WithToolchain(InProcessEmitToolchain.Instance))
                 .AddDiagnoser(MemoryDiagnoser.Default);
 
             BenchmarkSwitcher.FromAssembly(typeof(Program).Assembly).Run(args, doNotRequireSlnToRunBenchmarks);