--- a/src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RequestRoutingOptions.cs
+++ b/src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RequestRoutingOptions.cs
@@ -2,12 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System;
-using System.ComponentModel.DataAnnotations;
 
 namespace Microsoft.Extensions.Http.Resilience.Routing.Internal;
 
 internal class RequestRoutingOptions
 {
-    [Required]
     public Func<RequestRoutingStrategy>? RoutingStrategyProvider { get; set; }
 }