--- a/src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RoutingStrategyBuilder.cs
+++ b/src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RoutingStrategyBuilder.cs
@@ -2,21 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Options.Validation;
 
 namespace Microsoft.Extensions.Http.Resilience.Routing.Internal;
 
-internal sealed record RoutingStrategyBuilder : IRoutingStrategyBuilder
-{
-    public RoutingStrategyBuilder(string name, IServiceCollection services)
-    {
-        Name = name;
-        Services = services;
-
-        _ = Services.AddValidatedOptions<RequestRoutingOptions, RequestRoutingStrategyOptionsValidator>(name);
-    }
-
-    public string Name { get; }
-
-    public IServiceCollection Services { get; }
-}
+internal sealed record RoutingStrategyBuilder(string Name, IServiceCollection Services) : IRoutingStrategyBuilder;