… src/Libraries/Microsoft.Extensions.Http.Resilience/Internal/HttpRequestMessageExtensions.cs
… src/Libraries/Microsoft.Extensions.Http.Resilience/Routing/Internal/RequestRoutingStrategyOptionsValidator.cs
… test/Libraries/Microsoft.Extensions.Http.Resilience.Tests/Routing/RoutingResilienceStrategyTests.cs
… test/Libraries/Microsoft.Extensions.Http.Resilience.Tests/Routing/RoutingStrategyTest.cs